+++ src/main/java/carracing/Car.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 
 public class Car {
