+++ src/main/java/carracing/Game.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 
 import java.util.stream.IntStream;
