+++ src/main/java/carracing/GameRule.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 
 public class GameRule {
