+++ src/main/java/carracing/MovableStrategy.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 public interface MovableStrategy<T> {
