+++ src/main/java/carracing/OutputInterface.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 public interface OutputInterface {
     void showCurrentStatus(MoveResult moveResult);
