+++ src/main/java/carracing/RacingCondition.java
@@ -1,4 +1,4 @@
-package car_racing;
+package carracing;
 
 public class RacingCondition {
     private Integer carCount;
